Fix config flag order and lib, env changes
Continue to `make` and `make install` separately, the different sapi's
will race otherwise. This hopefully gets rid of the need for deparallel.
Build zip extension shared to fix quirky compiling in the main program.
Build tidy after the main install to avoid problems with system lib
linking.
Rework pecl, config adjustments. Installs to a directory in the prefix
that is symlinked to a custom directory in HOMEBREW_PREFIX. Patches
default config to support the separate extension dirs.